@@ -1,61 +1,82 @@
 {
   "name": "@stoplight/spectral",
-  "version": "0.1.0",
-  "description": "An object linting framework",
-  "author": "Stoplight, Inc. (support@stoplight.io)",
-  "license": "MIT",
-  "main": "lib/index.js",
-  "types": "lib/index.d.ts",
-  "files": [
-    "lib"
+  "version": "0.0.0",
+  "description": "A flexible object linter with out of the box support for OpenAPI v2 and v3.",
+  "keywords": [
+    "json linter",
+    "linter",
+    "json validator",
+    "validator",
+    "OpenAPI",
+    "Swagger",
+    "schema",
+    "API"
   ],
+  "sideEffects": false,
+  "homepage": "https://github.com/stoplightio/spectral",
+  "bugs": "https://github.com/stoplightio/spectral/issues",
+  "author": "Stoplight <support@stoplight.io>",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/stoplightio/spectral"
+  },
+  "license": "Apache-2.0",
+  "main": "src/index.ts",
+  "files": [
+    "**/*"
+  ],
+  "engines": {
+    "node": ">=8.3.0"
+  },
   "scripts": {
-    "build": "tsc -p tsconfig.json",
-    "format": "prettier --write \"src/**/*.{ts}\"",
-    "lint": "tslint src/**/*.{ts}",
-    "prepublishOnly": "yarn build",
-    "test": "jest",
-    "test.cover": "jest --coverage",
-    "test.watch": "jest --watch"
+    "build": "sl-scripts build",
+    "build.docs": "sl-scripts build:typedoc",
+    "commit": "git-cz",
+    "lint": "sl-scripts lint",
+    "lint.fix": "yarn lint --fix",
+    "release": "sl-scripts release",
+    "release.docs": "sl-scripts release:docs",
+    "release.dryRun": "sl-scripts release --dry-run --debug",
+    "test": "sl-scripts test",
+    "test.prod": "yarn lint && yarn test --coverage --maxWorkers=2",
+    "test.update": "yarn test --updateSnapshot",
+    "test.watch": "yarn test --watch"
   },
   "dependencies": {
-    "ajv": "6.5.x",
+    "@stoplight/types": "3.x.x",
+    "ajv": "6.x.x",
     "jsonpath": "git://github.com/stoplightio/jsonpath.git#fe90d42",
-    "lodash.get": "4.4.x",
-    "lodash.merge": "4.6.x",
-    "should": "13.2.x"
+    "lodash": ">=4.17.5"
   },
   "devDependencies": {
-    "@types/jest": "23.3.x",
+    "@stoplight/scripts": "3.1.0",
     "@types/jsonpath": "0.2.x",
-    "@types/lodash.get": "4.4.x",
-    "@types/lodash.merge": "4.6.x",
-    "@types/node": "10.12.x",
-    "jest": "23.6.x",
-    "prettier": "1.14.x",
-    "ts-jest": "23.10.x",
-    "tslib": "1.9.x",
-    "tslint": "5.11.x",
-    "tslint-config-prettier": "1.15.x",
-    "tslint-plugin-prettier": "2.0.x",
-    "typescript": "3.1.3"
+    "@types/lodash": "4.x.x",
+    "typescript": "3.2.2"
   },
-  "jest": {
-    "transform": {
-      "^.+\\.tsx?$": "ts-jest"
-    },
-    "testRegex": "(/__tests__/.*|(\\.|/)(test|spec))\\.(jsx?|tsx?)$",
-    "modulePathIgnorePatterns": [
-      "lib"
-    ],
-    "moduleFileExtensions": [
-      "ts",
-      "tsx",
-      "js",
-      "jsx",
-      "json",
-      "node"
-    ],
-    "testEnvironment": "node"
+  "lint-staged": {
+    "*.{ts,tsx}$": [
+      "yarn lint.fix",
+      "git add"
+    ]
+  },
+  "husky": {
+    "hooks": {
+      "commit-msg": "commitlint -E HUSKY_GIT_PARAMS",
+      "pre-commit": "lint-staged"
+    }
+  },
+  "config": {
+    "commitizen": {
+      "path": "node_modules/cz-conventional-changelog"
+    }
+  },
+  "commitlint": {
+    "extends": [
+      "@commitlint/config-conventional"
+    ]
+  },
+  "release": {
+    "extends": "@stoplight/scripts/release"
   }
 }
