@@ -1,11 +1,12 @@
-const merge = require('lodash.merge');
+import { oas2Functions, oas2Rules } from './oas2';
+import { oas3Functions, oas3Rules } from './oas3';
 
-import { IRuleset } from '../types';
-import { oas2Ruleset } from './oas2';
-import { oas3Ruleset } from './oas3';
+const merge = require('lodash/merge');
 
-export const defaultRuleset = (): IRuleset => {
-  return merge(oas2Ruleset(), oas3Ruleset(), {
-    name: 'all',
-  });
+export const defaultRules = () => {
+  return merge(oas2Rules(), oas3Rules());
+};
+
+export const defaultFunctions = () => {
+  return merge(oas2Functions(), oas3Functions());
 };
