@@ -1,9 +1,6 @@
 export const functions = {
   alphabetical: require('./alphabetical').alphabetical,
-  maxLength: require('./maxLength').maxLength,
-  notContain: require('./notContain').notContain,
-  notEndWith: require('./notEndWith').notEndWith,
-  or: require('./or').or,
+  length: require('./length').length,
   pattern: require('./pattern').pattern,
   schema: require('./schema').schema,
   truthy: require('./truthy').truthy,
