@@ -1,8 +1,6 @@
 import * as AJV from 'ajv';
 import * as jsonSpecv4 from 'ajv/lib/refs/json-schema-draft-04.json';
 
-import { IRuleFunction, IRuleOpts, IRuleResult, ISchemaRule, RuleSeverity } from '../types';
-
 const ajv = new AJV({
   meta: false,
   schemaId: 'auto',
@@ -14,14 +12,22 @@ ajv._opts.defaultMeta = jsonSpecv4.id;
 // @ts-ignore
 ajv._refs['http://json-schema.org/schema'] = 'http://json-schema.org/draft-04/schema';
 
-export const schema: IRuleFunction<ISchemaRule> = (opts: IRuleOpts<ISchemaRule>) => {
-  const results: IRuleResult[] = [];
+import { IFunction, IFunctionResult, ISchemaOptions } from '../types';
 
-  const { object, rule, meta } = opts;
-  const { schema: schemaObj } = rule.input;
+export const schema: IFunction<ISchemaOptions> = (targetVal, opts, paths) => {
+  const results: IFunctionResult[] = [];
+
+  if (!targetVal)
+    return [
+      {
+        message: `${paths.target ? paths.target.join('.') : 'property'} does not exist`,
+      },
+    ];
+
+  const { schema: schemaObj } = opts;
 
   // TODO: potential performance improvements (compile, etc)?
-  if (!ajv.validate(schemaObj, object) && ajv.errors) {
+  if (!ajv.validate(schemaObj, targetVal) && ajv.errors) {
     ajv.errors.forEach((e: AJV.ErrorObject) => {
       // @ts-ignore
       if (e.params && e.params.additionalProperty) {
@@ -30,11 +36,7 @@ export const schema: IRuleFunction<ISchemaRule> = (opts: IRuleOpts<ISchemaRule>)
       }
 
       results.push({
-        type: meta.rule.type,
-        path: meta.path.concat(e.dataPath.split('/').slice(1)),
-        name: meta.name,
-        summary: rule.summary,
-        severity: rule.severity ? rule.severity : RuleSeverity.ERROR,
+        path: paths.given.concat(e.dataPath.split('/').slice(1)),
         message: e.message ? e.message : '',
       });
     });
