@@ -1,41 +1,28 @@
-import { Spectral } from '../../../../../index';
-import { commonOasRuleset } from '../../../index';
+import { Spectral } from '../../../../../spectral';
+import { commonOasFunctions, commonOasRules } from '../../../../oas/index';
 
-const ruleset = commonOasRuleset();
+const ruleset = { functions: commonOasFunctions(), rules: commonOasRules() };
 
 describe('oasOp2xxResponse', () => {
-  const s = new Spectral({
-    rulesets: [
-      {
-        functions: ruleset.functions,
-        rules: {
-          oas2: {
-            'operation-2xx-response': Object.assign(
-              ruleset.rules['oas2|oas3']['operation-2xx-response'],
-              {
-                enabled: true,
-              }
-            ),
-          },
-        },
-      },
-    ],
+  const s = new Spectral();
+  s.addFunctions(ruleset.functions || {});
+  s.addRules({
+    'operation-2xx-response': Object.assign(ruleset.rules['operation-2xx-response'], {
+      enabled: true,
+    }),
   });
 
   test('validate a correct object', () => {
     const results = s.run({
-      spec: 'oas2',
-      target: {
-        paths: {
-          '/path1': {
-            get: {
-              responses: {
-                '200': {
-                  description: '',
-                  schema: {
-                    type: 'object',
-                    properties: {},
-                  },
+      paths: {
+        '/path1': {
+          get: {
+            responses: {
+              '200': {
+                description: '',
+                schema: {
+                  type: 'object',
+                  properties: {},
                 },
               },
             },
@@ -43,23 +30,20 @@ describe('oasOp2xxResponse', () => {
         },
       },
     });
-    expect(results.length).toEqual(0);
+    expect(results.results.length).toEqual(0);
   });
 
   test('return errors if missing 2xx', () => {
     const results = s.run({
-      spec: 'oas2',
-      target: {
-        paths: {
-          '/path1': {
-            get: {
-              responses: {
-                '400': {
-                  description: '',
-                  schema: {
-                    type: 'object',
-                    properties: {},
-                  },
+      paths: {
+        '/path1': {
+          get: {
+            responses: {
+              '400': {
+                description: '',
+                schema: {
+                  type: 'object',
+                  properties: {},
                 },
               },
             },
@@ -68,22 +52,19 @@ describe('oasOp2xxResponse', () => {
       },
     });
 
-    expect(results.length).toEqual(1);
+    expect(results.results.length).toEqual(1);
   });
 
   test('return errors if no responses', () => {
     const results = s.run({
-      spec: 'oas2',
-      target: {
-        paths: {
-          '/path1': {
-            get: {
-              responses: {},
-            },
+      paths: {
+        '/path1': {
+          get: {
+            responses: {},
           },
         },
       },
     });
-    expect(results.length).toEqual(1);
+    expect(results.results.length).toEqual(1);
   });
 });
